Version 0.4.0

+++ src/Controller/TaskController.php
@@ -36,6 +36,11 @@ class TaskController extends AbstractController
      */
     public function createAction(Request $request)
     {
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'Vous devez être authentifié pour créer une tâche');
+
+            return $this->redirectToRoute('task_list');
+        }
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
 
@@ -96,6 +101,9 @@ class TaskController extends AbstractController
      */
     public function deleteTaskAction(Task $task)
     {
+        // check for "edit" access: calls all voters
+        $this->denyAccessUnlessGranted('TASK_DELETE', $task);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
